# docs/swagger.yaml
basePath: /
definitions:
  AuthRequest:
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  AuthResponse:
    properties:
      expires_at:
        example: "2025-06-08T01:11:28Z"
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/UserInfo'
    type: object
...
  ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: Invalid request
        type: string
      message:
        example: The request is malformed
        type: string
    type: object
...
  ResearchProgressEvent:
    properties:
      progress:
        example: 25
        type: integer
      sources:
        example: 3
        type: integer
      status:
        enum:
        - processing
        - completed
        - error
        example: processing
        type: string
      step:
        example: Finding relevant sources...
        type: string
      timestamp:
        example: "2025-06-07T01:11:28Z"
        type: string
    type: object
...
  UserInfo:
    properties:
      email:
        example: user@example.com
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: John Doe
        type: string
      role:
        example: user
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: contact@deepresearch.ai
    name: DeepResearch Team
    url: http://www.deepresearch.ai
  description: AI-powered research platform API
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DeepResearch API
  version: 1.0.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            $ref: '#/definitions/AuthResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: User login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/AuthRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successful registration
          schema:
            $ref: '#/definitions/AuthResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: User registration
      tags:
      - auth
...
  /research/stream:
    get:
      description: Stream research progress in real-time using Server-Sent Events
      parameters:
      - description: Research query
        in: query
        name: query
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: Research progress stream
          schema:
            $ref: '#/definitions/ResearchProgressEvent'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Research streaming
      tags:
      - research
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token for API authentication (e.g., "Bearer {token}")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"